git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/trunk@2171 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/lluf/lluf_utils.ss/tests/LowLevelLogger_test/run_test.py
@@ -26,20 +26,20 @@
 from __future__ import print_function
 import subprocess, os, time, sys, shutil, re
 from threading import Thread
-
+import argparse
 try:
     # 3.x name
     from queue import Queue, Empty
 except ImportError:
     # 2.x name
     from Queue import Queue, Empty
-
+    
 def rmdir(directory):
     if os.path.exists(directory):
         try:
             shutil.rmtree(directory)
         except OSError:
-            print("Failed to remove directory, will retry")
+            print("Failed to remove directory",directory,", will retry")
             time.sleep(0.2)
             shutil.rmtree(directory)
 
@@ -103,8 +103,18 @@ def call_logger_control(args):
 if sys.platform == "win32":
     config_type = os.environ.get("CMAKE_CONFIG_TYPE")
     exe_path = config_type if config_type else ""
+    temp = os.environ.get("TEMP")
+    if temp is None:
+        temp = os.environ.get("TMP")
+    if temp is None:
+        print ("Failed to find temp dir!")
+        sys.exit(1)
+    logdir = os.path.join(temp, "safir_sdk_core", "log")
 else:
     exe_path = "."
+    logdir = os.path.join("/", "tmp", "safir_sdk_core", "log")
+
+print ("Logdir: ", logdir)
 
 lll_test = os.path.join(exe_path,"lll_test")
 
@@ -112,25 +122,27 @@ SAFIR_RUNTIME = os.environ.get("SAFIR_RUNTIME")
 
 logger_control = os.path.join(SAFIR_RUNTIME,"bin","logger_control")
 
-logdir = os.path.join(SAFIR_RUNTIME,"log")
-
 def logfilename(proc):
-    if sys.platform == "win32":
-        fn = "lll_test.exe-" + str(proc.pid) + ".txt"
-    else:
-        fn = "lll_test-" + str(proc.pid) + ".txt"
+    fn = "lll_test-" + str(proc.pid) + ".txt"
+    return os.path.join(logdir,fn)
 
-    return os.path.join(logdir,"Dob-LowLevelLog",fn)
+if len(sys.argv) != 2:
+    print ("Expect one argument")
+    sys.exit(1)
 
+configs_dir = sys.argv[1]
+if not os.path.isdir(configs_dir):
+    print ("arg is not a directory")
+    sys.exit(1)
+
+os.environ["SAFIR_TEST_CONFIG_OVERRIDE"] = os.path.join (configs_dir,"enabled_off")
 
 rmdir(logdir)
 
-os.mkdir(logdir)
-
-#Run the program that logs
+print ("Run the program that logs")
 p = LllProc()
 
-#check that log file only contains error texts
+print ("check that log file only contains error texts")
 found = False
 for i in range(100):
     data = p.logfile()
@@ -150,26 +162,25 @@ if data.find("Hello, World!") != -1 or data.find("Goodbye cruel world") != -1:
     p.kill()
     sys.exit(1)
 
-
-#turn logging on
+print ("turn logging on")
 res = call_logger_control(("5",))
 if res.find("Log level should now be 5") == -1:
     print("failed to turn logging on")
-    proc.kill()
+    p.kill()
     sys.exit(1)
 
-#let it log for a short while
+print ("let it log for a short while")
 p.wait_output(".*Hello, World!$")
 
-#turn logging off again
+print ("turn logging off again")
 res = call_logger_control(("0",))
 if res.find("Log level should now be 0") == -1:
     print("failed to turn logging off")
-    proc.kill()
+    p.kill()
     sys.exit(1)
 
 
-#check that log file isn't empty
+print ("check that log file isn't empty")
 data = p.logfile()
 if data.find("Hello, World!") == -1 or data.find("1234567890") == -1:
     print("failed to find expected log data in log file")
@@ -183,7 +194,7 @@ if data.find("Goodbye cruel world!") != -1:
     sys.exit(1)
 
 
-#sleep for a while and check that only errors are being logged
+print ("sleep for a while and check that only errors are being logged")
 time.sleep(0.5)
 
 newnum = p.logfile().count("Hello, World!")
@@ -193,7 +204,7 @@ if oldnum != newnum:
     sys.exit(1)
 
 
-#turn logging on again to level 9 and check that we get all logs
+print ("turn logging on again to level 9 and check that we get all logs")
 call_logger_control(("9",))
 p.wait_output(".*Goodbye cruel world!$")
 data = p.logfile()
@@ -205,25 +216,28 @@ if data.find("Hello, World!") == -1 or data.find("1234567890") == -1 or data.fin
 
 p.kill()
 
-#check that we get output on stdout
+print ("check that we get output on stdout")
 p = LllProc()
 if not p.wait_output(".*1234567890$"):
     print("lllerr does not output on stdout")
+    p.kill()
     sys.exit(1)
 p.kill()
 
-#check that the timestamps can be turned off
+print ("check that the timestamps can be turned off")
 p = LllProc()
 if not p.wait_output("^\[[0-9.:]*\] 1234567890$"):
     print("could not find line with timestamp")
+    p.kill()
     sys.exit(1)
 call_logger_control(("-t","0")) #turn timestamps off
 if not p.wait_output("^1234567890$"):
     print("could not find line without timestamp")
+    p.kill()
     sys.exit(1)
 p.kill()
 
-#check that flushing can be turned off
+print ("check that flushing can be turned off")
 p = LllProc()
 call_logger_control(("-i","0")) #turn flushing off
 time.sleep(0.1)
@@ -234,9 +248,10 @@ data = p2.output()
 if len(data) != 0:
     print("Flushing doesnt seem to be possible to turn off")
     print(data)
-    sys.exit(1)    
+    p.kill()
+    sys.exit(1)
 
-#turn flushing on and stdout off
+print ("turn flushing on and stdout off")
 call_logger_control(("-s", "0"))
 time.sleep(0.1)
 p2 = LllProc(False)
@@ -246,9 +261,10 @@ data = p2.output()
 if len(data) != 0:
     print("stdout logging doesnt seem to be possible to turn off")
     print(data)
+    p.kill()
     sys.exit(1)    
 
-#turn file off
+print ("turn file off")
 call_logger_control(("-f", "0"))
 time.sleep(0.1)
 p2 = LllProc()
@@ -258,40 +274,73 @@ data = p2.logfile()
 if len(data) != 0:
     print("file logging doesnt seem to be possible to turn off")
     print(data)
+    p.kill()
     sys.exit(1)    
 
 p.kill()
 
-time.sleep(0.1)
-rmdir(logdir)
+print ("check that disabling in inifile works")
+os.environ["SAFIR_TEST_CONFIG_OVERRIDE"] = os.path.join (configs_dir,"disabled")
 
-#check that we get lllerr output even if no logdir
-p = LllProc()
-if not p.wait_output(".*1234567890$"):
-    print("lllerr does not work without logdir")
+print ("Run the program that logs")
+p = LllProc(False)
+time.sleep(2.0)
+
+print ("check that we dont have a log file")
+try:
+    data = p.logfile()
+    print("have unexpected logfile")
+    sys.exit(1)
+except:
+    pass
+
+
+print ("try to turn logging on")
+res = call_logger_control(("5",))
+if res.find("LowLevelLogger is currently disabled") == -1:
+    print("Doesnt seem to be disabled!")
+    print(res)
+    p.kill()
+    sys.exit(1)
+
+time.sleep(2.0)
+data = p.output()
+if data.find("orld!") != -1:
+    print("there should not be any log output")
+    p.kill()
     sys.exit(1)
 
 p.kill()
 
-#test write permanent settings
-call_logger_control(("--create-logdir","-p","9"))
-p = LllProc()
-if not p.wait_output(".*Hello, World!$") and not p.wait_output(".*Goodbye cruel world!$"):
-    print("logger_control cannot write permanent settings")
-    sys.exit(1)
-p.kill()
+print ("check that enabling and turning on in inifile works")
+os.environ["SAFIR_TEST_CONFIG_OVERRIDE"] = os.path.join (configs_dir,"enabled_on")
 
-#test clear permanent settings
-call_logger_control(("-c",))
+print ("Run the program that logs")
 p = LllProc()
-p.wait_output(".*1234567890$")
-p.wait_output(".*1234567890$")
+
+print ("check that we get some logs")
+p.wait_output(".*Goodbye cruel world!$")
+
 data = p.logfile()
-if data.find("Hello, World!") != -1 or data.find("Goodbye cruel world!") != -1:
-    print("logger_control cannot write permanent settings")
+if data.find("Goodbye") == -1 or data.find("Hello") == -1 or data.find("1234567890") == -1:
+    print("failed to find all the expected stuff in the log file")
     sys.exit(1)
+
+print ("turn logging off")
+res = call_logger_control(("0",))
+if res.find("Log level should now be 0") == -1:
+    print("failed to turn logging off")
+    proc.kill()
+    sys.exit(1)
+
+print ("clear output and check that new logs are only error logs")
+p.output()
+time.sleep(0.5)
+data = p.output()
+if data.find("orld!") != -1:
+    print ("failed to turn off ini enabled")
+    sys.exit(1)
+
 p.kill()
-
-
 print("success")
 sys.exit(0)
